+++ http/createTransfer.go
@@ -10,7 +10,7 @@ import (
 
 type TransferRequest struct {
 	AccountDestinationId string `json:"account_Destination_Id"`
-	Amount               int64  `json:"amount"`
+	Amount               int    `json:"amount"`
 }
 
 func (s Server) CreateTransfer(w http.ResponseWriter, r *http.Request) {
